conditional install conda

+++ backend/models/pyramidflow/deploy.sh
@@ -7,15 +7,20 @@ cd Pyramid-Flow
 cp /home/ubuntu/text2vid-viewer/backend/models/pyramidflow/inference.py /home/ubuntu/Pyramid-Flow/inference.py
 
 # create env using conda
-wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O /tmp/miniconda.sh && \
-    bash /tmp/miniconda.sh -b -p $HOME/miniconda3 && \
-    rm /tmp/miniconda.sh
-export PATH="$HOME/miniconda3/bin:$PATH"
-conda init
-source ~/.bashrc
+if [ -f "$HOME/miniconda3/etc/profile.d/conda.sh" ]; then
+    . "$HOME/miniconda3/etc/profile.d/conda.sh"
+else
+    wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O /tmp/miniconda.sh && \
+        bash /tmp/miniconda.sh -b -p $HOME/miniconda3 && \
+        rm /tmp/miniconda.sh
+    export PATH="$HOME/miniconda3/bin:$PATH"
+    conda init
+    source ~/.bashrc
+    # Add a delay to ensure the environment is set properly
+    sleep 2  # Sleep for 2 seconds to allow the environment to refresh
+fi
 
-# Add a delay to ensure the environment is set properly
-sleep 2  # Sleep for 2 seconds to allow the environment to refresh
+
 
 conda create -n pyramid python=3.8.10 -y
 conda activate pyramid
